Move full-up tab test to test_mainwindow and minor cleanup
